change number typing

+++ gateway/sds_gateway/api_methods/utils/metadata_schemas.py
@@ -221,7 +221,7 @@ rh_capture_index_mapping = {
         "type": "float",
     },
     "altitude": {
-        "type": "float",
+        "type": "double",
     },
     "mac_address": {
         "type": "keyword",
